% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_point.R
\name{best_point}
\alias{best_point}
\title{Best Point Finder}
\usage{
best_point(c, se, sp)
}
\arguments{
\item{c}{c_rank scores}

\item{se}{sensitivity}

\item{sp}{specificity}
}
\value{
A list containing youden index and closest topleft points
}
\description{
Finding best point of ROC curve according to youden and closest topleft method.
}
