% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{body}
\alias{body}
\title{body part of main user interface}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
body
}
\description{
body part of main user interface
}
\keyword{datasets}
