% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_missings}
\alias{check_missings}
\title{Check Missing Values}
\usage{
check_missings(data_file)
}
\arguments{
\item{data_file}{dataset for checking its missing values}
}
\value{
returns a list containing total count of missing values, variables with low missing values,
a dataset with less than 10\% missing values, variables vector with less than 10 percent missing values,
variables vector with high missing values, missing patterns.
}
\description{
check missing values percentage and their patterins
}
