% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{complete_case}
\alias{complete_case}
\title{Complete Case Dataset}
\usage{
complete_case(low_miss_dataframe)
}
\arguments{
\item{low_miss_dataframe}{returned dataset with less than ten percent missing values from check_missing function, ie. missing_data function third value}
}
\value{
a list containing complete case dataset and included variables in complete case
}
\description{
creating a dataset with no missing values
}
