% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{data_generator}
\alias{data_generator}
\title{Fake data generator}
\usage{
data_generator(type = "standard", rows_num = 12, event = T, time = T)
}
\arguments{
\item{type}{determines the type of fake data as "standard" for data which has same
class and number of variables, "more" which has more variable other than main data,
"less" which at least lacks one variable compared to standard form
"mismatch", which has different class of some variabls compared to standard form.}

\item{rows_num}{number of observations of fake data}

\item{event}{whether event variable should be added, default is True}

\item{time}{whether time variable should be added, default is True}
}
\value{
fake data as data frame
}
\description{
To generate fake data for testing
}
\details{
by generating fake data with proper and improper variables,
we test uploding files and models functions.
}
