% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dataset_modification}
\alias{dataset_modification}
\title{Dataset Modifier}
\usage{
dataset_modification(dataset)
}
\arguments{
\item{dataset}{dataset read by read_data function}
}
\value{
modified dataset
}
\description{
module to peform some preprocessing steps automatically on new dataset privided by read_data function
}
