% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_predict.R
\name{final_predict}
\alias{final_predict}
\title{Prediction of each observation according to best points found by youden and closest topleft method}
\usage{
final_predict(predict_table, best_scores)
}
\arguments{
\item{predict_table}{predicted table of crank scores of observations}

\item{best_scores}{result of survROC function}
}
\description{
Prediction of each observation according to best points found by youden and closest topleft method
}
