% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{first_year_MACE}
\alias{first_year_MACE}
\title{First Year MACE}
\usage{
first_year_MACE(dataset)
}
\arguments{
\item{dataset}{dataset to be provided.}
}
\value{
dataset with MACE and censor time modfied for less than 365.25 days.
}
\description{
creating data for one year survival, by correcting event for time less than a year ie 365.25 days.
}
