% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{Loading trained models on original study}
\usage{
load_model(input)
}
\arguments{
\item{input}{input of shinyserver}
}
\description{
Loading trained models on original study
}
