% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_function.R
\name{loadingFunc}
\alias{loadingFunc}
\title{showing progressbar during loading a function}
\usage{
loadingFunc(message = "Loading ...")
}
\arguments{
\item{message}{message to display in progressbar}
}
\description{
showing progressbar during loading a function
}
