% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_prediction.R
\name{manual_prediction}
\alias{manual_prediction}
\title{Manual Predictor}
\usage{
manual_prediction(input)
}
\arguments{
\item{input}{input set by Shiny}
}
\value{
a dataframe containing user provided features of a patient
}
\description{
Prediction function for manually provided features
}
