% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{missing_impute}
\alias{missing_impute}
\title{mode and median missing imputation}
\usage{
missing_impute(dataset)
}
\arguments{
\item{dataset}{dataset to be imputed}
}
\value{
imputed dataset
}
\description{
replacing missing values by mode for categorical variable and median for numerical variables
}
