% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_scores.R
\name{predict_scores}
\alias{predict_scores}
\title{Main predicting tool of data by saved pre-trained machine learning models}
\usage{
predict_scores(data, model)
}
\arguments{
\item{data}{provided dataset we want to predict}

\item{model}{saved model to be used as prediction model}
}
\description{
Main predicting tool of data by saved pre-trained machine learning models
}
