% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_stage}
\alias{preprocess_stage}
\title{Data Preprocessor}
\usage{
preprocess_stage(
  url = "https://raw.github.com/hedayatbehnam/primace/main/R/data/dataset.csv"
)
}
\arguments{
\item{url}{url of new_dataset file in github located in R/data/new_dataset.RDS. directory}
}
\value{
uploaded new dataset
}
\description{
module to preprocess data automatically
}
\details{
when new dataset is available, this module automatically check its validity and
initiates preprocess as machine learning operationals.
}
