% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_variables}
\alias{raw_variables}
\title{raw dataset variable names}
\format{
A vector of raw dataset variables names
}
\source{
source dataset
}
\usage{
data(test_set)
}
\description{
raw dataset variable names
}
\references{
\url{primace.aikadeh.com}
}
\author{
Behnam Hedayat \email{dr.hedayatb@gmail.com}
}
\keyword{raw_variables}
