% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactiveVal_output.R
\name{reactiveVal_output}
\alias{reactiveVal_output}
\title{for each process one indicator of reactiveValue is created eg. perfMetrics with different
values when is non proessessed or complete to change text in each section before and after process}
\usage{
reactiveVal_output(name, value, content, output)
}
\arguments{
\item{name}{name of variable}

\item{value}{specific value of name input}

\item{content}{check content of output}

\item{output}{output set of Shiny}
}
\description{
for each process one indicator of reactiveValue is created eg. perfMetrics with different
values when is non proessessed or complete to change text in each section before and after process
}
