% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_data}
\alias{read_data}
\title{New Data Reader}
\usage{
read_data(
  url = "https://raw.github.com/hedayatbehnam/primace/main/R/datasets/dataset.RDS"
)
}
\arguments{
\item{url}{url of new_dataset file in github located in R/data/new_dataset.RDS. directory}
}
\value{
uploaded new dataset
}
\description{
module to read new data automatically from github repo
}
\details{
when new dataset is available, this module automatically reads it as entry point for machine learning operationals.
}
