% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{The Shiny App Server.}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{input set by Shiny.}

\item{output}{output set by Shiny.}

\item{session}{session set by Shiny.}
}
\description{
An app to predict first year survival following primary percutaneous coronary intervention.
}
\details{
This app can be used to predict and assess performance of user's provided datasets,
based of containing target variables, it also provides manual prediction tool to predict a patient,
according to its features selected online.
}
