% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{sidebar}
\alias{sidebar}
\title{sidebar of main user interface}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
sidebar
}
\description{
sidebar of main user interface
}
\keyword{datasets}
