% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{style_ref}
\alias{style_ref}
\title{style of html}
\usage{
style_ref(tags, htmlParam)
}
\arguments{
\item{tags}{tags for general use for various element}

\item{htmlParam}{which html parameters to referes styles to}
}
\description{
style of html
}
