% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survROC.R
\name{survROC}
\alias{survROC}
\title{survROC.}
\usage{
survROC(
  Stime,
  status,
  marker,
  entry = NULL,
  predict.time,
  cut.values = NULL,
  method = "NNE",
  lambda = NULL,
  span = NULL,
  window = "symmetric"
)
}
\arguments{
\item{Stime}{event or censor times}

\item{status}{status at event or censor time}

\item{marker}{crank scores}

\item{entry}{entry of survial time}

\item{predict.time}{time to predict}

\item{cut.values}{predefined cutoff of crank scores}

\item{method}{method to make analysis, either NNE or KM}

\item{lambda}{defining lambda value}

\item{span}{defining span value for NNE method}

\item{window}{defining window}
}
\value{
A list containing cut-off values, true positive rate, false positive rate,
specificity, prediction time, area under the curve at specified prediction time,
cumulative survival.
}
\description{
modified function of survivalROC package to return specificity as well.
}
\details{
It is used for prediction of survival according to time of interest.
}
