% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_set.R
\docType{data}
\name{test_set}
\alias{test_set}
\title{test_set Dataset}
\format{
A data frame containig example test set of main dataset
}
\source{
source dataset
}
\usage{
data(test_set)
}
\description{
test_set Dataset
}
\references{
\url{primace.aikadeh.com}
}
\author{
Behnam Hedayat \email{dr.hedayatb@gmail.com}
}
\keyword{test_set}
