% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varnames.R
\docType{data}
\name{varnames}
\alias{varnames}
\title{varnames Dataset}
\format{
A data frame containig variable names with their respective types and levels
}
\usage{
data(varnames)
}
\description{
varnames Dataset
}
\references{
\url{primace.aikadeh.com}
}
\author{
Behnam Hedayat \email{dr.hedayatb@gmail.com}
}
\keyword{varnames}
