% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_predict.R
\name{final_predict}
\alias{final_predict}
\title{Final Prediction}
\usage{
final_predict(predict_table, best_scores, target)
}
\arguments{
\item{predict_table}{predicted table of crank scores of observations}

\item{best_scores}{result of survROC function}

\item{target}{whether time and status variables are available}
}
\value{
predicted table according to best cut-off points of youden index and closest top left point
}
\description{
Prediction of each observation according to best points found by youden and closest topleft method.
}
