% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{Model Loader}
\usage{
load_model(
  input,
  rf_model = "models/rfsrc.learner_mod.RDS",
  xgb_model = "models/xgboost.learner_mod.RDS"
)
}
\arguments{
\item{input}{input of shinyserver}

\item{rf_model}{path to survival random forest model}

\item{xgb_model}{path to xgboost model}
}
\value{
loaded models in rds format
}
\description{
Loading trained models on original study
}
