% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_perf_calc.R
\name{max_perf_calc}
\alias{max_perf_calc}
\title{Performance Calculator}
\usage{
max_perf_calc(predict_table)
}
\arguments{
\item{predict_table}{predicted table of outcomes of each observation}
}
\value{
table of confusion matrix as dataframe
}
\description{
creating confusion matrix to show best performance result.
}
