% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{sidebar}
\alias{sidebar}
\title{User Interface}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
sidebar
}
\description{
Main user interface of primace app
}
\details{
Using shinydashboard package to create a dashboard like tabbed UI
sidebar of main user interface
}
\keyword{datasets}
