% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survROC.R
\name{survROC}
\alias{survROC}
\title{survivalROC modified function to return specificity as well}
\usage{
survROC(
  Stime,
  status,
  marker,
  entry = NULL,
  predict.time,
  cut.values = NULL,
  method = "NNE",
  lambda = NULL,
  span = NULL,
  window = "symmetric"
)
}
\arguments{
\item{Stime}{event or censor times}

\item{status}{status at event or censor time}

\item{marker}{crank scores}

\item{entry}{entry of survial time}

\item{predict.time}{time to predict}

\item{cut.values}{predefined cutoff of crank scores}

\item{method}{method to make analysis, either NNE or KM}

\item{lambda}{defining lambda value}

\item{span}{defining span value for NNE method}

\item{window}{defining window}
}
\description{
survivalROC modified function to return specificity as well
}
