% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{upload_data}
\alias{upload_data}
\title{function to upload data files by user}
\usage{
upload_data(input)
}
\arguments{
\item{input}{input set if Shiny}
}
\description{
function to upload data files by user
}
